Add files via upload
# Reaction Data AI Model

This project demonstrates the end-to-end process of building a machine learning model to predict reaction yields based on data extracted from `.pbtxt` files. The steps involved are:

1. **Data Parsing and Preprocessing**: Extracting and preparing data from `.pbtxt` files for model training.
2. **Model Training**: Creating and training a neural network model using TensorFlow/Keras.
3. **Model Evaluation**: Evaluating the trained model's performance on a test dataset.

## Project Structure


## Installation

Ensure you have Python installed. You will also need the following Python packages:

- `tensorflow`
- `scikit-learn`
- `pandas`
- `numpy`

You can install the required packages using `pip`:

```bash
pip install tensorflow scikit-learn pandas numpy

##LICENSE
*MIT License*

Copyright (c) [year] [your name]

Permission is hereby granted, free of charge, to any person obtaining a copy
of this software and associated documentation files (the "Software"), to deal
in the Software without restriction, including without limitation the rights
to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all
copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
SOFTWARE.